Remake `run_example` so it uses a driver library
