+++ src/CMakeLists.txt
@@ -24,8 +24,17 @@ add_library(WAVEFUNCTION
 add_library(PROJECTED_STATE
     ProjectedState.cpp
 )
+add_library(SPIN_MODEL
+    SpinModel.cpp
+)
 add_library(VMC_RUNNER
     VariationalMonteCarlo.cpp
+)
+add_library(VMC_DRIVER
+    VMCDriver.cpp
+)
+add_library(VMC_RESULTS
+    VMCResults.cpp
 )
 add_library(VMC_NUMERICAL
     vmc_numerical.cpp
@@ -45,4 +54,18 @@ set_target_properties(${TARGET_NAME} PROPERTIES COMPILE_FLAGS "${COMPILE_FLAGS}"
 set_target_properties(${TARGET_NAME} PROPERTIES LINK_FLAGS "${LINK_FLAGS}")
 add_custom_target(cpu-gpu DEPENDS ${TARGET_NAME})
 
-target_link_libraries(${TARGET_NAME} VMC_IO VMC_TYPE LATTICE MEAN_FIELD_ANSATZ PROJECTED_STATE WAVEFUNCTION VMC_RUNNER VMC_NUMERICAL $<LINK_ONLY:MKL::MKL>)
+target_link_libraries(
+    ${TARGET_NAME}
+    VMC_DRIVER
+    VMC_IO
+    VMC_TYPE
+    LATTICE
+    MEAN_FIELD_ANSATZ
+    PROJECTED_STATE
+    WAVEFUNCTION
+    SPIN_MODEL
+    VMC_RUNNER
+    VMC_NUMERICAL
+    VMC_RESULTS
+    $<LINK_ONLY:MKL::MKL>
+)
