+++ src/VMCResults.cpp
@@ -1,0 +1,4 @@
+#ifndef _USE_MATH_DEFINES
+#define _USE_MATH_DEFINES
+#endif // !_USE_MATH_DEFINES
+#include "mkl.h"
